Mod: Refactor constants and utils folder structure: put them into a new folder named: core
