+++ src/core/constants/app.constants.ts
@@ -59,7 +59,7 @@ export const DEV_AUTH_LOGOUT_URL = `${DEV_SERVER_URL}${AUTH_LOGOUT}`;
 export const PRD_AUTH_LOGOUT_URL = `${PRD_SERVER_URL}${AUTH_LOGOUT}`;
 
 /* Search params */
-export const ATOM_SEARCH_LIMIT = 6;
+export const ATOM_SEARCH_LIMIT = 9;
 export const ATOM_SEARCH_ORDER_BY_DEFAULT = 'likes';
 export const ATOM_SEARCH_ORDER = 'DESC';
