+++ src/core/utils/functionsUtil.js
