+++ src/core/utils/functionsUtil.ts
