+++ src/core/utils/interfaces.js
