+++ src/models/atom.model.ts
@@ -89,7 +89,7 @@ SequelizeStatic.Model<IAtomInstance, IAtomAttributes> {
             },
             views: {
                 type: dataTypes.INTEGER,
-                defaultValue: 0 // TODO: Validar al correr los migration files, que si se asigne este valor por defecto
+                defaultValue: 0
             },
             likes: {
                 type: dataTypes.INTEGER,
