+++ src/schema/atom/atom.query.js
@@ -4,7 +4,8 @@ Object.defineProperty(exports, "__esModule", { value: true });
 /*           DEPENDENCIES           */
 /************************************/
 const index_1 = require("./../../models/index");
-const pagination_1 = require("./../../utils/pagination");
+const pagination_1 = require("./../../core/utils/pagination");
+const appConfig = require("./../../core/constants/app.constants");
 // TODO: Agregar un mensaje descriptivo, y mover a un lugar adecuado
 function buildQueryFilter(isPrivate = false, atomCategoryId, text) {
     // Init Filter
@@ -23,38 +24,6 @@ function buildQueryFilter(isPrivate = false, atomCategoryId, text) {
         };
     }
     return queryFilter;
-}
-// TODO: Agregar un mensaje descriptivo, y mover a un lugar adecuado
-function buildPaginationQuery(before, after, desc, paginationField, primaryKeyField, paginationFieldIsNonId) {
-    const decodedBefore = !!before ? pagination_1.pagination.decodeCursor(before) : null;
-    const decodedAfter = !!after ? pagination_1.pagination.decodeCursor(after) : null;
-    // If is before (previous) = FALSE, if not TRUE
-    const cursorOrderIsDesc = before ? !desc : desc;
-    const cursorOrderOperator = cursorOrderIsDesc ? '$lt' : '$gt';
-    // VARIABLES
-    let paginationQuery;
-    let order = [
-        cursorOrderIsDesc ? [paginationField, 'DESC'] : paginationField,
-        ...(paginationFieldIsNonId ? [primaryKeyField] : []),
-    ];
-    if (before) {
-        paginationQuery = pagination_1.pagination.getPaginationQuery(decodedBefore, cursorOrderOperator, paginationField, primaryKeyField);
-        /* FIXME: #67 - Rompe cuando paginationFieldIsNonId es false, es decir
-            cuando quiero organizar por 'created_at' */
-        order = [
-            paginationField,
-            // ...(paginationFieldIsNonId ? [primaryKeyField, 'DESC'] : []),
-            paginationFieldIsNonId ? [primaryKeyField, 'DESC'] : '',
-        ];
-    }
-    else if (after) {
-        paginationQuery = pagination_1.pagination.getPaginationQuery(decodedAfter, cursorOrderOperator, paginationField, primaryKeyField);
-        order = [
-            cursorOrderIsDesc ? [paginationField, 'DESC'] : paginationField,
-            ...(paginationFieldIsNonId ? [primaryKeyField] : []),
-        ];
-    }
-    return { paginationQuery, order };
 }
 /**************************************/
 /*         ATOM QUERY TYPEDEF         */
@@ -122,8 +91,7 @@ exports.resolver = {
          * @param {Int} limit - limit number of results returned
          * @returns {Array<IAtom>} Atoms list
          */
-        // TODO: Crear un archivo de constantes como en el FE, para almacenar 12
-        allAtoms(parent, { limit = 12 }) {
+        allAtoms(parent, { limit = appConfig.ATOM_SEARCH_LIMIT }) {
             return index_1.models.Atom.findAll({
                 limit,
                 where: {
@@ -141,8 +109,7 @@ exports.resolver = {
          * @param {number} limit - limit number of results returned
          * @returns {Array<Atom>} Atoms List of a specific category (Buttons, Inputs, Labels, etc.)
          */
-        // TODO: Crear un archivo de constantes como en el FE, para almacenar 12
-        atomsByCategory(parent, { filter, limit = 12 }) {
+        atomsByCategory(parent, { filter, limit = appConfig.ATOM_SEARCH_LIMIT }) {
             return index_1.models.Atom.findAll({
                 limit,
                 where: {
@@ -156,72 +123,6 @@ exports.resolver = {
          * @method Method searchAtoms
          * @public
          * @param {any} parent - TODO: Investigar un poco más estos parametros
-         * @param {IAtomQueryArgs} args - destructuring: filter, limit, sortBy
-         * @param {IAtomFilterArgs} filter - a set of filters
-         * @param {String} sortBy - sort list by a passed parameter
-         * @param {number} limit - limit number of results returned
-         * @returns {Array<Atom>} Atoms List based on a filter parameters:
-         * e.g category, user's input text
-         */
-        // TODO: Crear un archivo de constantes como en el FE, para almacenar 'created_at' y 12
-        /*searchAtoms(
-            parent: any,
-            {
-                filter,
-                sortBy = 'created_at',
-                limit = 12
-            }: IAtomQueryArgs) {
-
-            // Init Filter
-            let queryFilter: IQueryFilters = {
-                active: true,
-                private: filter.private || false
-            };
-
-            // Add 'atomCategoryId' filter if it exists or is different from 0
-            if (filter.atomCategoryId && filter.atomCategoryId !== 0) {
-                queryFilter.atomCategoryId = filter.atomCategoryId;
-            }
-
-            // Add 'name' filter if 'text' exists
-            if (filter.text) {
-                queryFilter.name = {
-                    $like: `%${filter.text}%`
-                };
-            }
-
-            // Get all Atoms based on query args
-            // TODO: Crear un archivo de constantes como en el FE, para almacenar 'DESC'
-            return models.Atom.findAndCountAll({
-                where: queryFilter,
-                order: [[sortBy, 'DESC']],
-                limit
-            }).then((atoms) => {
-                const edges = atoms.rows.map((atom) => ({
-                    cursor: Buffer.from(atom.dataValues.id.toString()).toString('base64'),
-                    node: atom
-                }));
-
-                return {
-                    edges,
-                    count: atoms.count,
-                    pageInfo: {
-                        hasNextPage() {
-                            return true;
-                        },
-                        hasPreviousPage() {
-                            return true;
-                        }
-                    }
-                };
-            });
-
-        },*/
-        /**
-         * @desc Get Atoms by an user's input text (including category filter)
-         * @method Method searchAtoms
-         * @public
-         * @param {any} parent - TODO: Investigar un poco más estos parametros
          * @param {IAtomPaginationArgs} pagination - include: first, last, before, and after parameters
          * @param {IAtomQueryArgs} args - destructuring: filter, limit, sortBy
          * @param {IAtomFilterArgs} filter - a set of filters
@@ -229,7 +130,7 @@ exports.resolver = {
          * @param {number} limit - limit number of results returned
          * @returns {Array<Atom>} Atoms List based on a pagination params
          */
-        searchAtoms(parent, { filter = {}, sortBy = 'likes', pagination = {} }) {
+        searchAtoms(parent, { filter = {}, sortBy = appConfig.ATOM_SEARCH_ORDER_BY_DEFAULT, pagination = {} }) {
             // VARIABLES
             let { first, after, last, before } = pagination;
             let { isPrivate = false, atomCategoryId, text } = filter;
@@ -254,7 +155,7 @@ exports.resolver = {
             }
             where = Object.assign({}, where, filterQuery);
             // Build pagination query
-            let { paginationQuery, order } = buildPaginationQuery(before, after, desc, paginationField, primaryKeyField, paginationFieldIsNonId);
+            let { paginationQuery, order } = pagination_1.pagination.buildPaginationQuery(before, after, desc, paginationField, primaryKeyField, paginationFieldIsNonId);
             /* TODO: Si quito el 'any' me da error de type, ya que WhereOption del model
              no acepta: $and */
             const whereQuery = paginationQuery ? { $and: [paginationQuery, where] } : where;
