+++ src/schema/atom/atom.query.ts
@@ -3,7 +3,9 @@
 /************************************/
 import { models, sequelize } from './../../models/index';
 import { Buffer } from 'buffer';
-import { pagination as paginationUtils }  from './../../utils/pagination';
+import { pagination as paginationUtils }  from './../../core/utils/pagination';
+import * as appConfig from './../../core/constants/app.constants';
+
 
 // TODO: Agregar un mensaje descriptivo, y mover a un lugar adecuado
 function buildQueryFilter(isPrivate: boolean = false, atomCategoryId: number, text: string): IQueryFilters {
@@ -27,65 +29,6 @@ function buildQueryFilter(isPrivate: boolean = false, atomCategoryId: number, te
     }
 
     return queryFilter;
-}
-
-// TODO: Agregar un mensaje descriptivo, y mover a un lugar adecuado
-function buildPaginationQuery(
-    before: string, 
-    after: string, 
-    desc: boolean, 
-    paginationField: string,
-    primaryKeyField: string,
-    paginationFieldIsNonId: boolean
-) {
-
-    const decodedBefore = !!before ? paginationUtils.decodeCursor(before) : null;
-    const decodedAfter = !!after ? paginationUtils.decodeCursor(after) : null;
-    // If is before (previous) = FALSE, if not TRUE
-    const cursorOrderIsDesc = before ? !desc : desc;
-    const cursorOrderOperator = cursorOrderIsDesc ? '$lt' : '$gt';    
-
-    // VARIABLES
-    let paginationQuery;
-    let order: Array<any> = [
-        cursorOrderIsDesc ? [paginationField, 'DESC'] : paginationField,
-        ...(paginationFieldIsNonId ? [primaryKeyField] : []),
-    ];
-
-    if (before) {
-        
-        paginationQuery = paginationUtils.getPaginationQuery(
-            decodedBefore, 
-            cursorOrderOperator, 
-            paginationField, 
-            primaryKeyField
-        );
-
-        /* FIXME: #67 - Rompe cuando paginationFieldIsNonId es false, es decir 
-            cuando quiero organizar por 'created_at' */
-        order = [
-            paginationField,
-            // ...(paginationFieldIsNonId ? [primaryKeyField, 'DESC'] : []),
-            paginationFieldIsNonId ? [primaryKeyField, 'DESC'] : '',
-        ];
-
-    } else if (after) {
-
-        paginationQuery = paginationUtils.getPaginationQuery(
-            decodedAfter, 
-            cursorOrderOperator, 
-            paginationField, 
-            primaryKeyField
-        );
-
-        order = [
-            cursorOrderIsDesc ? [paginationField, 'DESC'] : paginationField,
-            ...(paginationFieldIsNonId ? [primaryKeyField] : []),
-        ];
-
-    }
-
-    return {paginationQuery, order};
 }
 
 
@@ -211,8 +154,7 @@ export const resolver = {
          * @param {Int} limit - limit number of results returned
          * @returns {Array<IAtom>} Atoms list
          */
-        // TODO: Crear un archivo de constantes como en el FE, para almacenar 12
-        allAtoms(parent: any, { limit = 12 }: IAtomQueryArgs) {
+        allAtoms(parent: any, { limit = appConfig.ATOM_SEARCH_LIMIT }: IAtomQueryArgs) {
             return models.Atom.findAll({
                 limit,
                 where: {
@@ -232,8 +174,7 @@ export const resolver = {
          * @param {number} limit - limit number of results returned
          * @returns {Array<Atom>} Atoms List of a specific category (Buttons, Inputs, Labels, etc.)
          */
-        // TODO: Crear un archivo de constantes como en el FE, para almacenar 12
-        atomsByCategory(parent: any, { filter, limit = 12 }: IAtomQueryArgs) {
+        atomsByCategory(parent: any, { filter, limit = appConfig.ATOM_SEARCH_LIMIT }: IAtomQueryArgs) {
             return models.Atom.findAll({
                 limit,
                 where: {
@@ -242,75 +183,6 @@ export const resolver = {
                 }
             });
         },
-
-        
-        /**
-         * @desc Get Atoms by an user's input text (including category filter)
-         * @method Method searchAtoms
-         * @public
-         * @param {any} parent - TODO: Investigar un poco más estos parametros
-         * @param {IAtomQueryArgs} args - destructuring: filter, limit, sortBy
-         * @param {IAtomFilterArgs} filter - a set of filters
-         * @param {String} sortBy - sort list by a passed parameter
-         * @param {number} limit - limit number of results returned
-         * @returns {Array<Atom>} Atoms List based on a filter parameters: 
-         * e.g category, user's input text
-         */
-        // TODO: Crear un archivo de constantes como en el FE, para almacenar 'created_at' y 12
-        /*searchAtoms(
-            parent: any, 
-            {
-                filter, 
-                sortBy = 'created_at', 
-                limit = 12
-            }: IAtomQueryArgs) {
-
-            // Init Filter
-            let queryFilter: IQueryFilters = {
-                active: true,
-                private: filter.private || false
-            };
-
-            // Add 'atomCategoryId' filter if it exists or is different from 0
-            if (filter.atomCategoryId && filter.atomCategoryId !== 0) {
-                queryFilter.atomCategoryId = filter.atomCategoryId;
-            }
-
-            // Add 'name' filter if 'text' exists
-            if (filter.text) {
-                queryFilter.name = {
-                    $like: `%${filter.text}%`
-                };
-            }
-
-            // Get all Atoms based on query args
-            // TODO: Crear un archivo de constantes como en el FE, para almacenar 'DESC'
-            return models.Atom.findAndCountAll({
-                where: queryFilter,
-                order: [[sortBy, 'DESC']],
-                limit
-            }).then((atoms) => {
-                const edges = atoms.rows.map((atom) => ({
-                    cursor: Buffer.from(atom.dataValues.id.toString()).toString('base64'),
-                    node: atom
-                }));
-
-                return {
-                    edges,
-                    count: atoms.count,
-                    pageInfo: {
-                        hasNextPage() {
-                            return true;
-                        },
-                        hasPreviousPage() {
-                            return true;
-                        }
-                    }
-                };
-            });
-
-        },*/
-
 
 
         /**
@@ -327,7 +199,7 @@ export const resolver = {
          */
         searchAtoms(parent: any, { 
             filter = <IAtomFilterArgs> {}, 
-            sortBy = 'likes', 
+            sortBy = appConfig.ATOM_SEARCH_ORDER_BY_DEFAULT, 
             pagination = <IAtomPaginationArgs> {}
         }: IAtomQueryArgs) {
 
@@ -360,7 +232,7 @@ export const resolver = {
             where = Object.assign({}, where, filterQuery);
             
             // Build pagination query
-            let { paginationQuery, order } = buildPaginationQuery(
+            let { paginationQuery, order } = paginationUtils.buildPaginationQuery(
                 before,
                 after,
                 desc,
